Reorganize other .adoc files to member-api.adoc
