--- module-api/src/docs/asciidoc/member/admin-get-members.adoc
@@ -1,25 +1,0 @@
-= Find Members By Admin
-:toc: left
-:toclevels: 4
-
-:source-highlighter: highlightjs
-:root: ../../../../module-api/src/docs/asciidoc
-
-== Request
-
-=== CURL:
-include::{snippets}/member/admin-get-members/curl-request.adoc[]
-
-=== HTTPie:
-include::{snippets}/member/admin-get-members/httpie-request.adoc[]
-
-=== Request Parameters:
-include::{snippets}/member/admin-get-members/request-body.adoc[]
-
-=== Request HTTP Example:
-include::{snippets}/member/admin-get-members/http-request.adoc[]
-
-
-== Response
-=== Response HTTP Example:
-include::{snippets}/member/admin-get-members/http-response.adoc[]
