--- module-api/src/docs/asciidoc/member/get-member.adoc
@@ -1,26 +1,0 @@
-= Find Specific Member
-:toc: left
-:toclevels: 4
-
-:doctype: book
-:icons: font
-:source-highlighter: highlightjs
-:root: ../../../../module-api/src/docs/asciidoc
-
-== Request
-
-=== CURL:
-include::{snippets}/member/get-member/curl-request.adoc[]
-
-=== HTTPie:
-include::{snippets}/member/get-member/httpie-request.adoc[]
-
-=== Request Parameters:
-include::{snippets}/member/get-member/request-body.adoc[]
-
-=== Request HTTP Example:
-include::{snippets}/member/get-member/http-request.adoc[]
-
-== Response
-=== Response HTTP Example:
-include::{snippets}/member/get-member/http-response.adoc[]
