--- module-api/src/docs/asciidoc/member/modify-member.adoc
@@ -1,22 +1,0 @@
-= Modify Member Info
-:toc: left
-:toclevels: 4
-:source-highlighter: highlightjs
-:root: ../../../../module-api/src/docs/asciidoc
-
-== Request
-=== CURL:
-include::{snippets}/member/modify-member/curl-request.adoc[]
-
-=== HTTPie:
-include::{snippets}/member/modify-member/httpie-request.adoc[]
-
-=== Request Parameters:
-include::{snippets}/member/modify-member/request-body.adoc[]
-
-=== Request HTTP Example:
-include::{snippets}/member/modify-member/http-request.adoc[]
-
-== Response
-=== Response HTTP Example:
-include::{snippets}/member/modify-member/http-response.adoc[]
