--- module-api/src/docs/asciidoc/member/restore-member.adoc
@@ -1,22 +1,0 @@
-= Restore Sign Out
-:toc: left
-:toclevels: 4
-:source-highlighter: highlightjs
-:root: ../../../../module-api/src/docs/asciidoc
-
-== Request
-=== CURL:
-include::{snippets}/member/restore-member/curl-request.adoc[]
-
-=== HTTPie:
-include::{snippets}/member/restore-member/httpie-request.adoc[]
-
-=== Request Parameters:
-include::{snippets}/member/restore-member/request-body.adoc[]
-
-=== Request HTTP Example:
-include::{snippets}/member/restore-member/http-request.adoc[]
-
-== Response
-=== Response HTTP Example:
-include::{snippets}/member/restore-member/http-response.adoc[]
