+++ module-api/src/test/java/com/woomoolmarket/ModuleApiApplicationTest.java
@@ -1,13 +1,5 @@
 package com.woomoolmarket;
 
-import com.woomoolmarket.common.RestDocsConfiguration;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.context.annotation.Import;
-
-@AutoConfigureMockMvc
-@AutoConfigureRestDocs
-@Import(RestDocsConfiguration.class)
 class ModuleApiApplicationTest {
 
 }
