+++ module-api/src/test/java/com/woomoolmarket/config/ApiDocumentationConfig.java
@@ -15,16 +15,19 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.WebApplicationContext;
 
 @Transactional
 @SpringBootTest
 @AutoConfigureMockMvc
-@AutoConfigureRestDocs
+@AutoConfigureRestDocs(uriScheme = "https", uriHost = "localhost", uriPort = 8443)
 @Import(RestDocsConfiguration.class)
 public class ApiDocumentationConfig {
 
     @Autowired
     protected MockMvc mockMvc;
+    @Autowired
+    protected WebApplicationContext context;
     @Autowired
     protected ObjectMapper objectMapper;
     @Autowired
@@ -39,5 +42,4 @@ public class ApiDocumentationConfig {
     protected CartService cartService;
     @Autowired
     protected ProductService productService;
-
 }
