+++ module-api/src/test/java/com/woomoolmarket/controller/board/BoardControllerTest.java
@@ -1,19 +1,5 @@
 package com.woomoolmarket.controller.board;
 
-import com.woomoolmarket.config.ApiDocumentationConfig;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+class BoardControllerTest {
 
-
-class BoardControllerTest extends ApiDocumentationConfig {
-
-    @BeforeEach
-    void initialize() {
-        em.createNativeQuery("ALTER TABLE BOARD ALTER COLUMN `board_id` RESTART WITH 1").executeUpdate();
-    }
-
-    @Test
-    void configTest() {
-        System.out.println("yaho");
-    }
 }
