+++ module-api/src/test/java/com/woomoolmarket/controller/order/OrderControllerDocumentation.java
@@ -1,0 +1,12 @@
+package com.woomoolmarket.controller.order;
+
+import com.woomoolmarket.config.ApiDocumentationConfig;
+import org.junit.jupiter.api.BeforeEach;
+
+class OrderControllerDocumentation extends ApiDocumentationConfig {
+
+    @BeforeEach
+    void initialize() throws Exception {
+        em.createNativeQuery("ALTER TABLE ORDERS ALTER COLUMN `order_id` RESTART WITH 1").executeUpdate();
+    }
+}
