+++ module-api/src/test/resources/org/springframework/restdocs/templates/asciidoctor/request-fields.snippet
@@ -1,0 +1,12 @@
+|===
+|필드|타입|필수값|설명|제한
+
+{{#fields}}
+|{{path}}
+|{{type}}
+|{{^optional}}true{{/optional}}
+|{{description}}
+|{{#constraint}}{{constraint}}{{/constraint}}
+
+{{/fields}}
+|===
