+++ module-core/src/main/java/com/woomoolmarket/domain/member/entity/Address.java
@@ -1,6 +1,7 @@
 package com.woomoolmarket.domain.member.entity;
 
 import javax.persistence.Embeddable;
+import javax.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,7 +14,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Address {
 
+    @Pattern(regexp = "^[\\w]{2,24}$")
     private String city;
+    @Pattern(regexp = "^[\\w]{2,24}$")
     private String street;
+    @Pattern(regexp = "^[\\d]{5,6}$")
     private String zipcode;
 }
